requirements.txt
# CrewAI core
crewai>=0.90.0
crewai-tools>=0.20.0

# Database
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
pymysql>=1.0.0

# Data processing
pandas>=2.0.0

# Template engine
jinja2>=3.1.0

# Environment variables
python-dotenv>=1.0.0

# LLM
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
